change list to info
